+++ src/test/java/fr/france/outils/OutilsStringTest.java
@@ -40,4 +40,42 @@ class OutilsStringTest {
             "Mont-de-Marsan", "Loir-et-Cher", "Blois", "Loire", "Saint-Étienne", "Haute-Loire", "Le Puy-en-Velay",
             "Loire-Atlantique", "Nantes", "Loiret");
 
+    @ParameterizedTest(name = "la chaine de caractères {0} une fois son accentuation supprimé doit être égale à {1}")
+    @MethodSource("genererArgumentsPourtest_supprimerAccentuation")
+    void test_supprimerAccentuation(String input, String excepted) {
+        assertEquals(excepted, OutilsString.supprimerAccentuation(input));
+    }
+
+    private static Stream<Arguments> genererArgumentsPourtest_supprimerAccentuation() {
+        return Stream.of(//
+                Arguments.of("É", "E"), //
+                Arguments.of("1234567890", "1234567890"), //
+                Arguments.of("ô", "o"), //
+                Arguments.of("à", "a"), //
+                Arguments.of("è", "e"), //
+                Arguments.of("ç", "c"), //
+                Arguments.of("'", "'"), //
+                Arguments.of("-", "-"), //
+                Arguments.of(" ", " "), //
+                Arguments.of("é", "e"));
+    }
+
+    @ParameterizedTest(name = "la chaine de caractères {0} une fois formaté doit être égale à {1}")
+    @MethodSource("genererArgumentsPourtest_formater_1parametre")
+    void test_formater_1parametre(String input, String excepted) {
+        assertEquals(excepted, OutilsString.formater(input));
+    }
+
+    private static Stream<Arguments> genererArgumentsPourtest_formater_1parametre() {
+        return Stream.of(//
+                Arguments.of("", ""), //
+                Arguments.of("test", "test"), //
+                Arguments.of("D'Orge", "dorge"), //
+                Arguments.of("D Orge", "dorge"), //
+                Arguments.of("TeEsT", "teest"), //
+                Arguments.of("Téèçà", "teeca"), //
+                Arguments.of("ÉÈÇÀ", "eeca"), //
+                Arguments.of("test   t@°0245est", "testtest"));
+    }
+
 }
